+++ include/aligator/core/traj-opt-problem.hxx
@@ -1,9 +1,10 @@
 /// @file
-/// @copyright Copyright (C) 2022-2024 LAAS-CNRS, INRIA
+/// @copyright Copyright (C) 2022-2024 LAAS-CNRS, 2022-2025 INRIA
 /// @brief Template definitions header.
 #pragma once
 
 #include "aligator/core/traj-opt-problem.hpp"
+#include "aligator/core/solver-util.hpp"
 #include "aligator/utils/mpc-util.hpp"
 #include "aligator/tracy.hpp"
 
@@ -18,6 +19,7 @@ TrajOptProblemTpl<Scalar>::TrajOptProblemTpl(
     xyz::polymorphic<CostAbstract> term_cost)
     : init_constraint_(std::move(init_constraint)), stages_(stages),
       term_cost_(std::move(term_cost)), unone_(term_cost_->nu),
+      xs_init_strategy_(xs_default_init<Scalar>),
       init_cond_is_state_error_(checkInitCondIsStateError()) {
   unone_.setZero();
 }