+++ include/aligator/solvers/fddp/results.hpp
@@ -29,8 +29,7 @@ ResultsFDDPTpl<Scalar>::ResultsFDDPTpl(
   xs.resize(nsteps + 1);
   us.resize(nsteps);
 
-  xs_default_init(problem, xs);
-  us_default_init(problem, us);
+  problem.initializeSolution(xs, us);
 
   gains_.resize(nsteps);